refactor: DeckGenerator의 덱 생성 기능을 Deck의 생성자에서 담당하도록 변경
